feat: Implement EOD (End-of-Day) functionality with multi-SKU support
- Added EodViewModel to manage UI state for EOD entries in the mobile app.
- Created backend endpoint for closing EOD with multi-SKU data processing.
- Implemented validation for EOD entries, ensuring SKU and quantity integrity.
- Introduced utility functions for converting between colli and pezzi.
- Added unit tests for colli to pezzi conversion and EOD entry validation.

+++ android/app/src/main/java/com/sasu91/dosapp/data/api/dto/ExceptionDto.kt
@@ -23,7 +23,9 @@ data class ExceptionRequestDto(
     @SerializedName("date")             val date: String,         // YYYY-MM-DD
     @SerializedName("sku")              val sku: String,
     @SerializedName("event")            val event: String,
-    @SerializedName("qty")              val qty: Int,
+    /** Quantity: colli (decimal) for ADJUST/UNFULFILLED, pezzi (integer) for WASTE.
+     *  Server converts colli->pezzi using SKU.pack_size. */
+    @SerializedName("qty")              val qty: Double,
     @SerializedName("note")             val note: String = "",
     @SerializedName("client_event_id")  val clientEventId: String? = null,
 )
